+++ utils/auth-axios.js
@@ -1,16 +1,16 @@
-const axios = require('axios');
+const axios = require('axios')
 const environments = require('./environments')
 
-const verifyEnv = function(environment) {
+const verifyEnv = function (environment) {
   var env = environments[environment.toLowerCase()]
   if (!env) {
     throw new Error(`environment ${environment} does not exist`)
   }
   return env
-};
+}
 
-const buildAxiosWithEnvAndAuth = function(appkit, environment) {
-  var env = verifyEnv(environment);
+const buildAxiosWithEnvAndAuth = function (appkit, environment) {
+  var env = verifyEnv(environment)
   return axios.create({
     baseURL: env.url,
     headers: {